Move GetCurrentNamespace to resource package
...which is more specific than helpers.
