+++ pkg/helpers/helpers.go
@@ -1,7 +1,6 @@
 package helpers
 
 import (
-	"io/ioutil"
 	"os"
 )
 
@@ -32,15 +31,4 @@ func GetEnv(key, defaultValue string) string {
 	}
 
 	return value
-}
-
-// GetCurrentNamespace returns the current namesapce from file system,
-// if the namespace is not found, it returns the defaultNamespace.
-func GetCurrentNamespace(defaultNamespace string) string {
-	nsBytes, err := ioutil.ReadFile("/var/run/secrets/kubernetes.io/serviceaccount/namespace")
-	if err != nil {
-		return defaultNamespace
-	}
-
-	return string(nsBytes)
 }
